Move text concatenation code to new text.c file
The text concatenation code was interspersed with
other things in sh.c, and having it in its own
file makes the code cleaner.
